+++ lib/screens/projects/projects_page.dart
@@ -491,11 +491,12 @@ class _ProjectsPageState extends State<ProjectsPage> {
               icon: Icons.edit_outlined,
               onTap: () => _openEditProject(project),
             ),
-          DataCardAction(
-            icon: Icons.delete_outline,
-            color: AppColors.error,
-            onTap: _isDeleting ? () {} : () => _deleteProject(project),
-          ),
+          if (_canDeleteProjects)
+            DataCardAction(
+              icon: Icons.delete_outline,
+              color: AppColors.error,
+              onTap: _isDeleting ? () {} : () => _deleteProject(project),
+            ),
         ],
       ),
     );
